// src/models/sheet/Skills.ts
export type Skill = 
  |  'acrobatics' 
  | 'arcana' 
  | 'athletics' 
  | 'animalHandling'
  | 'performance' 
  | 'deception' 
  | 'stealth' 
  | 'history' 
  | 'intimidation' 
  | 'insight' 
  | 'investigation' 
  | 'medicine' 
  | 'nature' 
  | 'perception' 
  | 'persuasion' 
  | 'sleightOfHand' 
  | 'religion' 
  | 'survival'


export type SkillProficiency =
  | "none"
  | "proficient"
  | "expertise"